feat: merge api files
