+++ ui-vue3/src/api/service/service.ts
@@ -24,3 +24,19 @@ export const searchService = (params: any): Promise<any> => {
     params
   })
 }
+
+export const getServiceDetail = (params: any): Promise<any> => {
+  return request({
+    url: '/service/detail',
+    method: 'get',
+    params
+  })
+}
+
+export const getServiceDistribution = (params: any): Promise<any> => {
+  return request({
+    url: '/service/distribution',
+    method: 'get',
+    params
+  })
+}
